fix a bug where variants weren't handled properly in bulk card import
